requirements.txt
flask
azure-search-documents
azure-cognitiveservices-vision-computervision
azure-storage-blob
azure-cosmos
python-dotenv